Report the paths to the worst objects
It is frustrating to find out that there is, say, a gigantic blob in a
repository with no easy way to find out where it is in the
history. (Git's tooling for answering that question is almost
nonexistent.) So...

* Define a new class, `PathResolver`, that can compute the
  paths (i.e., the `git rev-parse` expression) that can be used to
  access an object with a specific OID. This class must be fed the
  objects and references in depth-first order, but that's OK because
  that's the order we need to process them to compute their expanded
  sizes.

* Change `Sizes` to record `*Path` pointers, rather than OIDs, for the
  big objects that it has found.

* If `Sizes` finds a new object that is bigger than the biggest that
  it has seen before, have it discard the (maybe only partly-computed)
  path of the old object and replace it with the new one.
  `PathResolver` knows how to free resources associated with the
  partly-computed paths, so basically the number of paths being
  computed at any time is at most one for each of the "largest" things
  that we want to track.

* Output the paths (still in crude format) to the JSON output.

This costs about 15% in runtime and hardly any additional memory
consumption.

Still to do: (1) Improve the format of the path output. (2) Emit
better output for commits that are not referred to directly by
references (perhaps using `git name-rev`). (3) Include the paths in
non-JSON output, and probably change where they are presented in JSON
format.