Add footnotes containing the paths to the tabular output
For the non-JSON output, add footnote after the table listing the
"biggest" objects referred to by the table, with bracketed citation
numbers in the table indicating the appropriate footnote. If an object
is "biggest" in more than one way, make the lines in the table refer
to the same footnote.