Systematically prefer to cite more recent objects
Given two objects that are the same size, or multiple trees/commits
that refer to the same large object, which one should we prefer to
cite in the footnotes? I think that new objects are more interesting
than old objects [1], because

* It is interesting to know whether the "bad" objects are still in
  use.

* Once you have a commit containing a "bad" object, it is easier to
  trace it backwards in time than forwards.

* New objects are more likely to be at the tips of branches and thus
  more likely to have a reference pointing at them.

The old code was somewhat inconsistent because blobs and trees are
processed in roughly reverse-chronological order whereas commits are
processed in roughly chronological order. Improve the situation by
implementing two pairs of `Count.AdjustMax()` methods,
`AdjustMaxIfNecessary()` and `AdjustMaxIfPossible()`, differing in the
obvious way if the two values that they are comparing are equal. By
choosing which variant to use in which situation, we can arrange to
preferentially pick the new of equal-sized objects.

Note that this still doesn't cause newer commits to be preferred in
the paths of trees and blobs.

[1] Note that there are also arguments that the old objects are more
    interesting, because they are more likely to indicate the commits
    where a large object was first introduced.