Support a minimum threshold for reporting statistics
Allow the user to specify a minimum threshold "level of concern" below
which statistics (and any corresponding footnotes) are not reported in
the tabular output. If a section has no interesting statistics,
suppress the headers of the section, too. And if the whole table has
no interesting statistics, just output

    No problems above the current threshold were found

The threshold is configured using the following options:

* `--threshold=<value>` — set the threshold to a numerical value, which
  is interpreted as the fewest number of stars that a statistic has
  to have for it to be displayed.

* `--verbose` — report all statistics, whether they are concerning or
  not.

* `--critical` — only report "critical" statistics; i.e., those which
  are displayed as exclamation points.

Set the default threshold to 1.