By default, just assume a tty
Use of cgo is a pain for cross-compilation, and also if the person
running the build doesn't have a C compiler installed. So make the
default version of `isatty.Isatty()` just return true. The only effect
of this change is that progress information is output by default.

If you want to really use `isatty()`, compile with

    make USE_ISATTY=t

or

    go build --tags isatty

Users can still control the output of progress information using
`--progress`/`--no-progress`.