NewRepository: don't join paths if `gitDir` is already absolute
If the output of `git -C path rev-parse --git-dir` is already an
absolute path (e.g., if `git-sizer` is run from within a subdirectory
of a working copy), then it is incorrect to call `filepath.Join()`
because that function doesn't treat absolute paths specially. So check
`filepath.IsAbs()` first, and only call `filepath.Join()` if the path
is not already absolute.

Fixes #10.