Don't consider a "maximum tag depth" of 1 to be concerning
Every tag has at least a tag depth of 1, so there's no need to flag it
as a reason for concern. This change is done a bit kludgily, by
changing the scale factor for this statistic from 1 to 1.001, but it
has the desired effect of subtracting one from the number of stars
computed for any reasonable tag depth.