ReferenceFilter: simplify how filters are combined
Instead of the special `IncludeExcludeFilter`, treat filters just as
sets, with the usual set operations (union, intersection, and
inversion) and build them up that way. Add a `Combiner` that knows how
to combine two `ReferenceFilter`s, and use it to implement "include"
vs. "exclude". Build special behavior into the combiners if the left
argument is `nil`, to get the old semantics that the "default" for an
unmentioned reference depends on whether the first directive is
"include" vs. "exclude".