Don't record unwalked references with the path resolver
Recording unwalked references with the path resolver means that the
footnote for an extreme object might use an unwalked references in the
"names" for the object, even though the object must necessarily also
be reachable from at least one reference that _was_ walked. Fix this
by not recording unwalked references with the path resolver.