Pipeline: prefer to report non-pipe errors to the caller
If a later stage of the pipeline fails, it is often the case that
earlier stages fail with some sort of pipe-related error (because
their `stdout` pipe gets closed). Those pipe errors are usually not
very helpful in figuring out the underlying problem. So if there are
any non-pipe errors, report the earliest one of those. If there are
only pipe errors, report the latest one of those.