Don't print the referecence count when explicit roots are provided
We printed out the reference count, even when explicit roots were
provided.
```
git-sizer master~50^{tree} 'HEAD@{1.week.ago}' -v
[TRUNCATED_OUTPUT]
| * References                 |           |                                |
|   * Count                    |   187     |                                |
|     * Ignored                |   187     |                                |
|                              |           |                                |
[TRUNCATED_OUTPUT]
```

This is not very useful, as the reference count depends on the list of references
extracted by running `git for-each-ref`.

This commit changes the output to not print the reference count when
explicit roots and no reference filters are provided to git-sizer