Tell Git that the bare repositories used in the tests are safe
As of Git v2.38.0, there is an option to prevent Git from accessing bare
repositories unless asked for explicitly (via `--git-dir` or `GIT_DIR`):
`safe.bareRepository`.

The tests of `git sizer`, however, assume that Git will access a bare
repository when the current directory points inside that repository.
This only works if `safe.bareRepository` indicates that this is safe.

Let's just override whatever value the system or user-wide Git config
has for that config setting, to ensure that the tests can pass.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>